elf: Add GNU_PROPERTY_NO_MEMORY_SEAL gnu property
On a glibc recent proposal [1] to add the Linux mseal support [2],
Mike Hommey raised that this feature might potentially break Firefox
on Linux.  Firefox bypasss the system loader by building with DT_RELR
support, and post-processed itself with a tool to both remove the
GLIBC_ABI_DT_RELR dependency and instrument the binaries to apply
the relocation themselves so they can deploy Firefox regardless if
loader supports DT_RELR (some more details at [3]).

To accomplish it, the instrumentation mimics the dynamic loader and
temporarily undoes the RELRO machine to be able to apply the
relocations, and redoes it afterward.  This is exactly what mseal
aims to prevent.

The GNU_PROPERTY_NO_MEMORY_SEAL gnu property is a way to mark such
objects to not be sealed by glibc.  The version 2 of glibc support
for memory sealing uses this new property [5].

[1] https://sourceware.org/pipermail/libc-alpha/2024-June/157359.html
[2] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=8be7258aad44b5e25977a98db136f677fa6f4370
[3] https://sourceware.org/pipermail/libc-alpha/2024-June/157668.html
[4] https://glandium.org/blog/?p=4297
[5] https://sourceware.org/pipermail/libc-alpha/2024-July/158804.html